Clean up a bunch of signed/unsigned comparison warnings
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@23120 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/include/k5-int.h
@@ -620,8 +620,8 @@ struct addrlist {
 	void (*freefn)(void *);
 	void *data;
     } *addrs;
-    int naddrs;
-    int space;
+    size_t naddrs;
+    size_t space;
 };
 #define ADDRLIST_INIT { 0, 0, 0 }
 extern void krb5int_free_addrlist (struct addrlist *);
